+++ src/main/java/app/repository/AcessorioRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.entity.Acessorio;
 
-public interface AcessorioRepository extends JpaRepository<Acessorio, Long> {
+public interface AcessorioRepository extends JpaRepository<Acessorio, Integer> {
 	
 	
 }
