+++ src/main/java/app/repository/CarroRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.entity.Carro;
 
-public interface CarroRepository extends JpaRepository<Carro, Long> {
+public interface CarroRepository extends JpaRepository<Carro, Integer> {
 
 	public List<Carro> findByNome(String nome);
